requirements.txt
requests
pytest
structlog
curlify
pyhamcrest